requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.0.6
django-autoslug==1.9.8
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-js-asset==2.0.0
Pillow==9.2.0
psycopg2==2.9.4
pycodestyle==2.9.1
sqlparse==0.4.2
toml==0.10.2
